move local address book to zChain
